refactor: relocate UsStateComboBox, table skeleton, ErrorBoundary, and BulkActionsModule
