+++ src/components/BankTransactions/BankTransactionsHeader.tsx
@@ -20,6 +20,7 @@ import { MonthPicker } from '@ui/datePickers/MonthPicker/MonthPicker'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { Toggle } from '@ui/Toggle/Toggle'
 import { Heading } from '@ui/Typography/Heading'
+import { BulkActionsModule } from '@blocks/BulkActionsModule/BulkActionsModule'
 import { Header } from '@blocks/layout/Container/Header'
 import { BankTransactionsBulkActions } from '@components/BankTransactions/BankTransactionsBulkActions/BankTransactionsBulkActions'
 import { BankTransactionsHeaderMenu, BankTransactionsHeaderMenuActions } from '@components/BankTransactions/BankTransactionsHeaderMenu'
@@ -28,7 +29,6 @@ import { RecordTransactionMenuButton } from '@components/BankTransactions/Record
 import { SelectedBankAccountsChip } from '@components/BankTransactions/SelectedBankAccountsChip/SelectedBankAccountsChip'
 import { TransactionsSearch } from '@components/BankTransactions/TransactionsSearch/TransactionsSearch'
 import { BankTransactionsActions } from '@components/BankTransactionsActions/BankTransactionsActions'
-import { BulkActionsModule } from '@components/BulkActionsModule/BulkActionsModule'
 import { SyncingComponent } from '@components/SyncingComponent/SyncingComponent'
 
 import './bankTransactionsHeader.scss'
