+++ src/components/blocks/BulkActionsModule/bulkActionsModule.scss
